+++ tools/build.py
@@ -103,6 +103,7 @@ APLICACIO = [
     # seguiment, document i dades
     "40-vista-seguiment.js",
     "50-document.js",
+    "51-document-edicio.js",
     "60-copies.js",
     "61-dades-io.js",
     "62-alumnat-io.js",